Fix Wireless Display bug for sample app
Updated the sample app to "properly"/gracefully handle Wireless Displays. Which
is achieved by adding MediaControlIntent.CATEGORY_LIVE_VIDEO to the
MediaRouteSelector. However, this may cause duplicate routes in the chooser
dialog in the future if the MediaRouter library or framework starts reporting
Secondary Displays or updates how Wireless Display routes are handled.

Closes SteadfastInnovation/SecondaryDisplayMediaRouteProvider#1