Skip default route when checking for duplicate displays
During testing, discovered that mirroring the screen using the Google
Home app caused the default route to contain the presentation display
that is being mirrored to. However, since most consuming applications
ignore the default route both in the route selector and when
choosing/configuring routes, that would cause the display being mirrored
to not show up as a route option (since we wouldn't be providing a route
for it).